Add batching functionality and corresponding tests
This commit introduces a new feature for partitioning collections into batches. This is achieved by creating new `Batch` and `Batches` classes with appropriate methods.

In particular, the `Batch` class represents a batch of elements while the `Batches` class provides methods for adding new batches and changing the size of existing batches. An extension method on `IEnumerable<T>` is also created to generate the Batches.

Additionally, tests are included to ensure correct functionality. This feature will be beneficial for memory efficient processing of large collections.